requirements.txt
sentence-transformers == 2.2.2
huggingface-hub<0.15
langchain>=1.0,<2.0
langchain-classic>=1.0,<2.0
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community>=0.4,<0.5
langchain-text-splitters>=1.0,<2.0
langchain-openai>=1.0,<2.0
langchain-experimental>=0.4,<0.5
-e .
